generator_app/requirements.txt
Flask==0.10.1
Flask-Assets==0.10
Flask-Babel==0.9
Flask-Bootstrap==3.3.0.1
Flask-Environments==0.1
Flask-Hashing==1.1
Flask-Login==0.2.9
Flask-Mail==0.9.1
Flask-Migrate==1.3.0
Flask-OAuth2-Login==0.0.9
Flask-OAuthlib==0.9.1
Flask-OpenID==1.2.4
Flask-Principal==0.4.0
Flask-SQLAlchemy==2.0
Flask-Script==2.0.5
Flask-Testing==0.4.2
Flask-WTF==0.10.3
Flask-WhooshAlchemy==0.56
Flask-Resize==0.8.0
PyYAML==3.12
osmapi==0.6.1
GeoAlchemy2==0.2.4
httplib2==0.9
psycopg2==2.5.4

